+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Show Ignored Recommendations (from Project ... menu)
 - Fix for Capacitor 5 migration with Cordova plugins that require jetifier
 - Fix for new projects where a number was used as part of the application name
+- Fix for Capacitor 5 starter projects to set the default Gradle JDK to 17
 
 ### Version 1.44